✨ feat(state): 状態用の構造体を作成
