+++ client/src/state/categories.rs
@@ -1,0 +1,5 @@
+pub struct CategoriesStateItem {
+    pub id:   i32,
+    pub name: String,
+    pub url:  String,
+}
